requirements.txt
amqp==5.0.6
asgiref==3.4.1
backcall==0.2.0
billiard==3.6.4.0
celery==5.2.1
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
decorator==5.1.0
Deprecated==1.2.13
Django==3.2.9
django-celery-beat==2.2.1
django-environ==0.8.1
django-timezone-field==4.2.1
djangorestframework==3.12.4
environ==1.0
ipython==7.29.0
jedi==0.18.0
kombu==5.2.2
matplotlib-inline==0.1.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.22
psycopg2-binary==2.9.2
ptyprocess==0.7.0
Pygments==2.10.0
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
redis==4.0.2
six==1.16.0
sqlparse==0.4.2
traitlets==5.1.1
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
